requirements.txt
fastapi
uvicorn[standard]
pyyaml
transformers
torch
peft
datasets
pytest

